feat(mongo.go): add MongoAdapter interface and MongoAdapterImpl struct
This commit adds a new file `mongo.go` which contains the definition of the `MongoAdapter` interface and the implementation struct `MongoAdapterImpl`. The `MongoAdapter` interface defines the methods for interacting with a MongoDB database, such as `InsertOne`, `UpdateOne`, `DeleteOne`, and `FindOne`. The `MongoAdapterImpl` struct implements these methods by delegating the calls to the underlying `MongoAdapter` instance.

This commit also includes the creation of a factory function `NewMongoAdapter` which creates a new instance of `MongoAdapterImpl` with the provided `MongoAdapter` instance.

The purpose of this change is to provide a flexible and modular way to interact with a MongoDB database by abstracting the implementation details behind the `MongoAdapter` interface. This allows for easier testing and swapping out different implementations of the adapter in the future.